refactor: split main page component and factor out subcomponents
